+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed 
 - [#42] Update Makefiles to 9.5.0
+- [#42] Update postgresql to 14.15
+- [#42] Add migration checker to post-upgrade.sh
+  - The migration checker finds potentially corrupt data in postgres
+  - Any found issues have to be fixed manually, see constraint_migration.md
 
 ## [v14.13-1] - 2024-11-13
 ### Changed